implement basic @Routed annotation idea #13
